fix: Externalize the more drop down menu customizability
